Code generation of Python and C++ sockets now working
Next need to add R. Then of course test the heck out of everything.

The file caching is getting concerning. I should be deleting these files
when I am done with them, but I will be replacing the whole file-based
system soon enough with RocksDB, so maybe the need is not so presisng.