Benchmark for C++/python map
Here are some benchmarks:

 +------------------+------------------+----------------+------------+
 | benchmark        | new              |  prior         | fold       |
 +------------------+------------------+----------------+------------+
 | cZeroBaseline    | 51.9   ± 3.6     | 43.9   ± 4.5   | -          |
 | pZeroBaseline    | 101.9  ± 5.9     | 82.1   ± 4.0   | -          |
 | pZeroFromForeign | 122.3  ± 65.0    | 89.0   ± 6.7   | -          |
 | pZeroToForeign   | 118.3  ± 62.0    | 90.8   ± 5.7   | -          |
 | rZeroBaseline    | 639.8  ± 79.6    | 311.5  ± 5.5   | --         |
 | rZeroFromForeign | 1016.5 ± 252.6   | 345.0  ± 11.9  | ---        |
 | rZeroToForeign   | 950.2  ± 366.1   | 334.6  ± 6.6   | ---        |
 +------------------+------------------+----------------+------------+
 | cTenBaseline     | 230.8  ± 85.2    | 98.2   ± 5.7   | --         |
 | pTenBaseline     | 202.0  ± 71.2    | 133.2  ± 4.9   | --         |
 | pTenFromForeign  | 4867.2 ± 605.9   | 2231.6 ± 49.2  | --         |
 | pTenToForeign    | 2059.2 ± 216.7   | 2041.2 ± 43.7  | =          |
 | rTenBaseline     | 2080.2 ± 320.2   | 1099.4 ± 34.0  | --         |
 | rTenFromForeign  | 4212.8 ± 51.0    | 4190.7 ± 418.3 | =          |
 | rTenToForeign    | 3858.1 ± 86.7    | 3528.7 ± 143.6 | =          |
 +------------------+------------------+----------------+------------+
 | rMarginalCost1   | 661.5  ± 8.3     | 328.2  ± 7.9   | -          |
 | rMarginalCost2   | 858.6  ± 10.0    | 637.0  ± 21.6  | -          |
 | rMarginalCost3   | 1050.6 ± 21.6    | 930.9  ± 55.3  | -          |
 | rMarginalCost4   | 1258.0 ± 15.3    | 1221.0 ± 38.4  | =          |
 | pMarginalCost1   | 103.8  ± 4.9     | 88.0   ± 4.4   | -          |
 | pMarginalCost2   | 110.6  ± 3.9     | 124.3  ± 4.1   | +          |
 | pMarginalCost3   | 113.1  ± 4.6     | 166.2  ± 4.9   | +          |
 | pMarginalCost4   | 118.3  ± 4.0     | 214.4  ± 10.7  | ++         |
 +------------------+------------------+----------------+------------+
 | pMap1000         | 4.6    ± 0.4     | 47.0   ± 5.1   | ++++++++++ |
 +------------------+------------------+----------------+------------+

The transition to sockets should accelerate cases were many foreign
calls are made. The pMap1000 is an extreme case of this, where a Python
function is called from a C++ loop 1000 times. The mean time for a
Python call from C for the new code is under 5ms, whereas it is ~50ms
for the old code.

Similarly, benchmarks for the marginal cost of an increasing number of
alternating language calls favors the socket version increasingly as the
depth increases.

For most other benchmarks, though, the performance is now worse. I could
dig into the reasons for this. Probably part of it is just the
complexity of extra libraries to import and extra functions in install.
In the more drastic cases, I suspect that large files are being copied
more than necessary.

Anyway, I'll leave deeper analysis for later. I suspect some of the
problems will disappear after I add the RocksDB.