Nexus and python pools working with new binary fmt
I've developed an complete binary package format for morloc.

My manjaro linux machine mysteriously crashed an hour ago. The system
booted just fine, but any attempt to log into my account just returned
me to the password entry screen. So I logged in through the shell
(Ctrl-Alt-F4 on my system), which is where I am now. I can still listen
to music through the CLI, as I always do, and my bash environment is
about the same as ever. The fonts are a bit screwy. There is really only
one irritating thing -- my CAPS-LOCK as reverted to being a CAPS-LOCK
rather than an ESC. So programming in vim is a pain. I also can't use
the internet, but I don't have a connection here anyway, so that is no
bother. If I can fix the CAPS issue, maybe I'll just always log in this
way?

Anyway, I was talking about my new binary format. I guess I should
describe it here (even though lack of ESC is irritating me).

All communication between proccesses is mediated by data sent in
packets. Each packet begins with a 32 byte header:

| 4 bytes  | 8 bytes | 8 bytes | 4 bytes | 8 bytes |
| magic    | version | command | offset  | length  |

 magic: The morloc packet identifier (6DF80707), pronounced mo-ding-ding
 version: 8 bytes of versioning info:
   a. 2 byte plain version
   b. 2 byte binary version
   c. 2 byte binary version flavor
   d. 2 byte mode
 command: an 8 byte sequence that describe what the packet does
 offset: the number of bytes after the header where the data starts
 length: the number of bytes in the data

So the entire packet size is 32 + offset + length.

Currently offset is always 0. But in the future I will use it to store
any metadata that does not fit in the header. What data is stored in
this region depends on the mode and command.

Currently there are 9 commands:

 1. DATA:    00 src fmt cmp enc 00 00 00
 2. CALL:    01 |    mid      | 00 00 00
 3. CALLRET: 02 res 00  00  00  00 00 00
 4. GET:     03 00  00  00  00  00 00 00
 5. GETRET:  04 res 00  00  00  00 00 00
 6. PUT:     05 00  00  00  00  00 00 00
 7. PUTRET:  06 res 00  00  00  00 00 00
 8. PING:    07 00  00  00  00  00 00 00
 9. PINGRET: 08 00  00  00  00  00 00 00

For DATA, the src, fmt, cmp, and enc variables respectively represent
the data source (raw data, file, database), the format (JSON, MsgPack,
etc), the compression format (e.g., gzip), and the encryption format.
Currently, I pass data only in JSON format with no compression or
encryption.

For CALL, the `mid` variable stores the manifold id as a 4 byte integer.
The arguments to the function are stored as a sequence of DATA packets
starting at index 32+offset.

In CALLRET, GETRET, and PUTRET, the `res` variable stores a PASS or FAIL
value that indicates whether the attached DATA packet is a result or an
error message.

The GET/GETRET and PUT/PUTRET commands are used to query the nexus for
data that is stores. The pools know nothing about how the nexus stores
info. They can send the nexus info with PUT and then they receive a
"key" to the info from the PUTRET return packet. This "key" can be used
to retrieve info from the nexus with a GET command. The nexus may store
this information in a key/value database of some form (I will probably
use SQLite in the first iteration).

Small arguments, with lengths less than 2^8 bytes, are currently passed
as raw data in the packet. Larger arguments are first stored by the
nexus (with a PUT command) and then retrieved when they are needed with
a GET command. This circumvents the need to pass large serialized
objects through potentially multiple foreign calls leading to multiple
copies.

So that is the rough idea, I'll elaborate later in the docs.

With this commit, the nexus and a single python command can communicate
small data (integers). This shows that the packets for nexus and the
python pools are being created correctly (at least the CALL, CALLRET,
PING, and PINGRET packets).

TODO:

 [ ] get nexus/C++ working
 [ ] get nexus/R working
 [ ] fix interop between all languages
 [ ] test for large data (where files are currently created)
 [ ] ensure all existing tests work
 [ ] clean up the code
 [ ] benchmark

After all this, I can next work on incorportating SQLite as a key/value
store. This will require working out PUT/PUTRET and GET/GETRET commands
and will require proper streaming of large data between the pools and
the nexus (where not all the data will fit in the buffers).

After getting SQLite working, I'll work on replacing JSON, and all the
ugly serialization libraries I've written, with MsgPack.

After that, I'll add a TCP option to the nexus, so that it can
communicate over a network, rather than only as a CLI program.

Then I will finally have the prototype of a real production system.