(All pass) Fix integer reading width in C++
I need a careful and systematic test of argument sizes.

The benchmark tests that return, for example, 1000 integers, are
currently failing. The reason is that the packets are longer than the
buffers and I haven't added support for this to the nexus.

I need to add medium (4097 - 65536) and long (65536+) packet length
tests to ensure large packet streaming and large data handling is
working correctly. And I need to be sure that all languages test read
and writing for these cases.

Eventually, I should also add some parameters to the compiler for
setting the buffer size (currently 4096) and the max packet size
(currently 65536).