Large interop working
Sockets are working correctly with large and medium data. There is still
a slight problem with the R implementation. I failed to find the bug in
the socketr.c package (assuming that is where the bug was) that was
causing R to fail on packets that exceeded the buffer size. So I just
increased the buffer size to something large enough to handle any
inter-process message. This is not as kludgy as it sounds, argument
packet size is limited to 2^16 bytes, anything larger than that will be
written to a file and passed by filename. A call packet can have many
argument data packets. But since arguments are positional, it is
unlikely to have more than a few of them. I've set the buffer size to
2^20 + 32, allowing 16 max-sized arguments plus the call packet header to
fit in the buffer. This is not great, and I will of course fix this bug
eventually, but it will be fine for now. If someone does surpass the
buffer size, they will get a friendly error message.