Add MessagePack handling to nexus and python
MessagePack format is much cleaner than JSON. There are no formatting
degeneracies or numeric fuzziness. For this reason, I'm planning on
purging JSON entirely from the morloc internals. However, I do still
need it for user input. So the nexus parses all incoming JSON data into
MessagePack. For files, if they are below the packet size (2^16 - 32
bytes) they are loaded into message data packets, otherwise a convert
them to MessagePack and write them as files to the same path as the JSON
file but with the MessagePack extension (".mpk"), then I pass them as
file data packets.

Python MessagePack conversions are quite tidy. Only tuples need special
handling. I am reserving the MessagePack extension number 0 for tuples.

R is proving difficult, though. The conversion from MessagePack is easy
enough (I just have to handle tuple extensions). For moving to
MessagePack, however, type annotations are required to resolve scalars
and vectors. The R package RcppMsgPack will pack a vector of length one
as an atomic element, but a vector of length 0 or 2+ will be packed as
an array. This is fine except in the edge case where what is intended to
be an array just so happens to have only one element. In this case, I
would need to explicitly recast the pseudo-scalar as a list to ensure it
is properly cast. But just by looking at the data, though, I can't know
what is intended to be a scalar, so I will need type annotations.

But recasting all the R types may be slow in some situations. For
example, in the type `[(Int,[Str])]`, I would need to loop through every
tuple in the list and cast the second element to a list if it contains
just one element.

What I may try is doing C++ next. Then skipping the RcppMsgPack library
and adapting the C++ code to handle R. I would still need to pass a
schema.