Made a single file mlcmpack header
This simplifies the code. But it does require a stupid hack to make gcc
compile a header into a shared object. The thing is that I only need the
shared object for the Python and R FFIs. For C++, I only need code, so a
single file solution is cleaner (I could use the shared object and the
header, but I've been having difficulty making that compile).

This isn't really a pretty solution. But it fixes things for now. I can
tidy it up later.