Fixed bad tests
Wow, the record-promotion-2 case was a trip. It failed in the
messagepack serialization with a terribly unhelpful assert fail in the C
library. As I dug in, I eventually found that "Alice" was being asigned
to the "age" field. Apparently this was being caused because I had an
incorrect definition of `snd` in python. I guess I had copied the
definition of `fst` and forgot to update the index, so `snd` was the
same as `fst`.

Then the old test was passing because I had lazily copied the observed
value over the expected value, hiding the incorrect definition of snd.