(58/429 fail) Fixed local-2
Sheesh ... what was smoking? I was treating a map like a record. Why did
this ever compile? Well, the typing was all correct, but the python code
was code wrong. And nothing in the code then caught the mistake. The new
code is much more strict. The error messages, however, are abysmal. I
need to fix that ...