Thoughts on Anything
At the heart of my msgpack library is the Anything struct. Every new
language I add will need to create a C FFI that maps native types to the
Anything struct. Then schema-based serialization can be conducted in a
language agnostic fashion.

```
typedef struct Anything {
    morloc_serial_type type;
    size_t size;
    char* key;
    union {
        char nil_val;
        bool bool_val;
        int int_val;
        double double_val;
        char* char_arr;
        bool* bool_arr;
        int* int_arr;
        double* float_arr;
        struct Anything** obj_arr;
    } data;
} Anything;
```

There are two problems with this struct. First, it is too big. It stores
a type (4 bytes), a size (8 bytes), the key (8 bytes) and the union (8
bytes). Second, it is too complex. I don't need to store the type, I
have a schema for that. Creation of the Anything structs is
schema-directed already, so no need to copy the schema into every
element. The key field is only needed for maps, so wastes space in most
fields. The size is only needed in arrays. The fields in the union are
too specific and have no good representation for boxed and unboxed
arrays. It is not easily extensible and requires heavy code duplication
to support all the specific cases. Why did I design it like this? Well,
I was loosely trying to mirror the types in the morloc_serial_type enum.
But that is no excuse.

Here is a better alternative:

```
typedef struct BoxedArray {
  size_t size;
  void** elements;
} BoxedData;

typedef struct UnboxedArray {
  size_t size;
  size_t width;
  void* elements;
} BoxedData;

typedef struct NamedArray {
  size_t size
  char** key;
  Anything2* val;
} BoxedData;

typedef struct Anything {
    union {
      int8_t i8,
      uint8_t u8,
      int16_t i16,
      uint16_t u16,
      int32_t i32,
      uint32_t u32,
      int64_t i64,
      uint64_t u64,
      char* cstring,
      BoxedArray* boxed_array,
      UnboxedArray* unboxed_array,
      void* object
    } data;
} Anything;
```

The idea here is for Anything to store abstract data by width with no
opinion about its type. So I remove the type annnotation and move all
size and key info into child structs. Now Anything has a width of 8
bytes. It can be easily overloaded with new types, such as datetime as a
u64 element. Schema information needn't be duplicated.

Tuples could be encoded as UnboxedArray types where width is 8 (the size
of an Anything struct) and size is the number of items in the tuple.

Moving to this representation will be a bit of a chore. There is a lot
of code to rewrite. But I think it will be worth it.

Another tricky consideration is memory management. Ideally, the Anything
objects should be created with no copying of memory. In some languages
this is possible. For example, R types are compatible with C under the
hood. I can directly set an R pointer to an double array to the elements
field in an UnboxedArray. No need for copying. This is not the case,
though, with native Python floats. So whether the C library or the
wrapper should free the memory may vary between languages.

One solution would be for the Anything object itself to store ownership
information. Whether we copy memory or directly pass a pointer the
Anything struct is a local decision made in the code that translates the
objects. So it makes sense for that code to be able to record its
decision directly in the object, rather than, for example, writing it
into the schema or enshrining it in some global policy outside the
codebase.

```
typedef struct BoxedArray {
  unsigned int local_spine : 1;
  unsigned int local_kids : 1;
  size_t size;
  void** elements;
} BoxedData;

typedef struct UnboxedArray {
  bool local;
  size_t size;
  size_t width;
  void* elements;
} BoxedData;

typedef struct NamedArray {
  unsigned int local_keys : 1;
  unsigned int  local_spine : 1;
  size_t size
  char** key;
  Anything2* val;
} BoxedData;

typedef struct CString {
  bool local;
  char* c_str;
} CString

typedef struct Object {
  bool local;
  void* obj;
} Object

typedef struct Anything {
    union {
      uint8_t u8,
      uint16_t u16,
      uint32_t u32,
      uint64_t u64,
      CString* cstring,
      BoxedArray* boxed_array,
      UnboxedArray* unboxed_array,
      Object* object
    } data;
} Anything;
```

With this, when we leave the mlcmpack library we can free the created
Anything object and recursively check whether every pointer is local and
should thus  be freed. Further, as we create the data, we can specify
the data we have allocated locally and we may vary our strategy between
languages.