(5/429) Fixed the string/list handling in R
There are random tests failing. The reason appears to be that libraries
are being rebuilt while tests are being run in parallel. The
SystemConfig.hs module builds shared libraries if they are missing. I
think a better approach would be to have a `morloc init` command that
builds all the required libraries and checks to ensure that everything
is as it should be. Having the compiler silently run these system
commands is perhaps not desirable.