(3/429 fail) Fixed handling of file data packets
To be clear, when a data packet contains data from a file source, the
file is encoded in the given format (say MessagePack), but the filename
is a simple string. I was previously encoding the filename in
MessagePack.

It might be worthwhile to add a STRING format type. I am currently
passing error messages in MessagePack format which only serves to make
them longer and more complicated.

Moving to a totally different topic, my temporary file handling is still
a mess. I chose not to fix it earlier since I was planning on replacing
the temporary file solutions with a database. However, I think the
overhead of sending data to the nexus, writing it to a database, and
retrieving it, is will not be lower than simple file IO. The database
may allow caching, which would speed things up. But not enough.

A better idea may be to just use memory mapped files. I would still need
to create intermediate files. These should be stored in the morloc tmp
folder. There they can be automatically cleaned or whatever (eventually
we may want to keep them to allow reentrency and such).