Place long running packets test last and est time
Now the long running tests packets-large and packets-interop are last.
This means we don't have to wait for them to run before catching easier
errors. Also they have a time estimate (~10s) so we now how long to wait
and can distinguish between infinite loops and normal behavior.

Speaking of which, if the pool servers fail silently the nexus will
continue listening for them forever. This failure should not happen
under normal circumstances since most errors are correctly caught. I
need to extend exception handling to neatly catch all cases. That would
of course be ideal. But I should also add in a ping every now and then
to check if the language pools are alive.

Another concern is the growing complexity of the runtime. The nexus
manages many language servers. Each language server spins up a process
for a each task. This is great if there a lot of tasks. But it could
cause small things to become to heavy. So far, this is only a serious
problem with R, where the extra 300ms is percetible to human users.