Create a basic shared memory library
This library exports the functions:

    void shinit(const char* shm_name, size_t shm_size);
    void shmalloc(size_t size);
    void shfree(void* ptr);
    void shcalloc(size_t nmemb, size_t size);
    void shrealloc(void* ptr, size_t size);
    void shclean();

These allow management of memory in a shared space. The memory model is
very rudimentary. It does not support defragmentation or resizing.
Defragmentation I can implement with move function specialized for the
`Anything` structure. I would loop through the pointers to the
`Anything` structures and move all their elements into contiguous memory
left-shifted as far as possible. If there is no memory shared between
`Anything` structures (which I think there will not be in my first
implementation), I can compact the memory further by removing all
internal headers and fitting the `Anything` space into a single block.

As for resizing, we can make a linked list of shared memory volumes.
When the first fills, we create the next (probably larger) volume. I'd
need a method for re-mmap-ing these. They could be described in the
header of the root volume and checked before memory is accessed. Maybe
via a `refresh` function.

Next I need to refactor the `Anything` generators in all languages to
use these new memory functions. This will be easy enough in R and C++,
but may require a rather complete rewrite in Python. I'll have to pass
the raw Python data as C objects over to a C library. I need to drop
down a level.