Import new mlcmpack.h and cppmpack.hpp files
These files were tested in the msgpack-parser library. The focus though
is on a much cleaner alternative to the awkwary Anything structure I
previously stored everything in. Now I just use void*. Yes, my structure
between data structure collapsed into the void.

This memory structure will be what is shared between languages in the
future shared memory pool. The between data has the following grammar:

    data : primitive
         | array data
         | tuple [data]
         | record [name]* [data]
    primitive : i8, i16, i32, i64, u8, u16, u32, u64, f32, f64

The primitives include signed and unsigned integers of various widths
and 32- and 64-bit floats. Booleans and nil are special cases of u8.

The array type is a 16-byte struct including an 8-byte size and an
8-byte pointer to the data.

An n-tuple is n fixed-size objects stored in contiguous memory. That is,
the tuple requires no extra space.

A record is a tuple. The keys to the record are stored separately in the
Schema object.

The Schema stores the data type, all parameters as child Schema objects,
the width of the type in bytes, the keys (as needed), and the offsets to
the beginnings of tuple children. The schema are required to map the
voidstar data to local structures.

So going forward, I am next going to rollback all the crappy work I did
on memory-mapped files. Then I am going to extend the new voidstar
method (that's the name I'm going with because it sounds cool) to all
other languages. This will be a bit of work. For Python, I will need to
replace my current ctypes trash with something written at a lower level
with the C Pythong API. Importantly, all memory management will be
delegated to mlcmapck.h. This is vital because eventually it will
allocate memory to the shared memory pool. All interactions with the
memory allocator will be through this one C file. Which most certainly
means it needs a better name.

My new implementation also opens up much more precise types. Finally I
can specify the integer and float widths. I'll need to rewrite all my
libraries to be compatible with this. Well, not really. The current
C/C++ `int` is i32, as is the R `integer` type. Python integers have
arbitrary width. R is a bit of a special duck here though, since R users
typically use doubles for everything, even places that are more
logically integers. It would be reasonable to store R integers as
doubles, and thus R "integers" would effectively be 52-bit (since
IEEE754 64bit ints include 1 sign bit, 11 exponent bits, and 52
precision bits).

After all this, I can then replace the memory management in `mlcmpack.h`
(or whatever future me names it) with a shared memory pool. This shared
memory allocator is already mostly written, but it will doubtless take
significant work to optimize and debug it.

Finally, I will need to add a few compiler options for managing memory.
These might include setting the initial pool size and the aggressiveness
of the garbage collector / defragmenter.