Add hidden debugging options to nexus
There are three options:

 -d <DIR> : use the persistent folder DIR rather than a temporary one
 --no-start : do start the language servers
 --leave-open : do not kill the servers and delete pipes on close

Together these allow the pools to be fired up independently of the
nexus. This is useful because I can open a pool with a debugger and
listen to messages coming in from the nexus.

I've found the GNU debugger (gdb) invaluable in diagnosing issues with
the message pack handling and binary conversions to and from voidstar.
To deal with forking, these GDB settings are helpful:

set follow-fork-mode child
set detach-on-fork off

I have been trying to test the new C++ code. And I walked through the
C++ pool, caught the message from the nexus and unpacked it. But the
result was not what I expected. Ends up that the nexus was sending over
packets with length of 1 and value of 0x00. Null values. I was expecting
i32 values. It looks like the problem is that the nexus is calling the
python message pack pack function which relies on the same shared
library as the C++ pool. But I haven't synced the Python pympack.py
code. So I guess I'll need to do that next.

Back to debugging, what is the future of morloc debugging? What should a
morloc debugger look like? One option would be to make a unified
debugger that communicates via pipes with all the pools. Execution would
pause at each step within a manifold and wait for a user command. The
user could query the local state as they like using compositions of a
set of atomic commands each language would need to implement.

An alternative idea is to use existing debuggers on just one language at
a time. This was my approach above with C++.

Of course, the two are not mutually exclusive.