Convert to single morloc.h file
This one morloc.h file currently contains message pack serialization,
morloc universal binary wrappers, and the new shared library code.
Eventually I will also move the packet and socket handling into this
header.

Now, should all this be in one giant C header file? The advantage of
this is that it simplifies builds across the board. The disadvantage is
that having everything in one file is ugly. Hmm, honestly, the argument
for lots of files is a bit weak. It improves organization, but it
requires an extra level of management to ensure files are loaded and
linked in the right way.

A better argument against the single-header solution is that it pollutes
the namespace. It would definitely be cleaner to have only the minimal
set of required functions exposed through a header and everything else
hidden in a source .c file.