Fix error propagation
Previously errors in a running function could cause the program to hang
when the failing packet was passed to another language. Also the
original error message from the failed function was lost. Now both
issues are fixed, though still the returning error message is not
formatted very well. In the future, I should turn this into a colorful
traceback and names of the actual functions called (rather than manifold
ids) and suggestions for where in the morloc code the problem arises.