Move demos to dediated examples repo
The flu case study was already included in the `examples` repo at
https://github.com/morloc-project/examples. I don't need to maintain
duplicates. I've also moved over the sequence analysis repo to that
folder. The morloc main repo should contain as little excess material as
possible to minimize the download size. There may be hundreds of
examples in the future, and many may have modest datasets associated.
They should live in their own place.

`morloc` main will contain a broad automated test suite and benchmarking
suite (soon to be automated). It should contain the minimum required for
testing in continuous integration. All these tests will be required to
pass for the latest version of morloc (and perhaps optionally for past
ones as well for backwards compatibility). The examples may be pegged to
older versions of the compiler and may have their own dedicated
containers.