Simplify `morloc init`
It was crashing previously in my vanilla Ubuntu container.
readCreateProcessWithExitCode failed in this container, complaining the
hGetContents couldn't parse one of its inupts. hGetContents is used
inside the readCreateProcessWithExitCode function to read STDOUT and
STDERR. One of those, for reasons I cann't fathom, is receiving
non-textual data. When I run the command myself STDERR is empty and
STDOUT is plain old ASCII. Not sure what is wrong. But I can replace the
readCreateProcessWithExitCode function with a simple callCommand
function. This does not intercept errors and doesn't allow me to make
pretty colors for failing cases, but at least it passes in the
container.