Fix error raised on NULL return
Empty values, such as those returned from functions run for their side
effects, like simple print statements that write to files, are already
handled correctly internally. The are stored as 0 bytes in voidstar
format. However, in the final checking in the nexus I was raising an
error if no content was returned, this led to an incorrect failing state
after successfully running the code. That is now fixed.

However, in the course of testing and fixing this code, I was reminded
of an old problem with morloc: how to handle STDERR and STDOUT. In the
past, I passed data between pools and the nexus through the STDOUT
channel. In order for this to work, I had to ensure that the channel was
not corrupted by any STDOUT written by the wrapped functions. So I
captured any STDOUT from these runs and stored it in the MorlocMonad for
possible later use. Same with STDERR. This behavior, though, hampers
debugging and is no longer necessary since I pass data over sockets.