Add specific numeric types
Extend the base types to include the following fixed width numeric types:

  Int8
  Int16
  Int32
  Int64
  UInt8 - bytes
  UInt16
  UInt32
  UInt64
  Float32
  Float64

This allows more strictness when desired and allows preservation of
numeric precision across languages. At least as far as the given
language allows. There are some caveats. We may need to add warnings
when passing data to languages that simply do not support the given
numeric type.

The `Int` and `Real` types are still defined. These will continue to
represent "ideal" integers and reals and languages will be free to do
with them as they please. True arbitrary sized ints and arbitrary
precision and size reals will need their own typeclass where they would
(probably) map to `[UInt8]` and `([UInt8], [UInt8])` respectively. Where
the integer is represented as a list of bytes (will need to be
consistent about endianess) and the real is a tuple of a big integer and
an exponent. Anyway, all that is 3rd party.

There is also the matter of 128 bit ints. Rust has them. For now, I
don't think we need them. There are not many cases I can think of where
a 128 bit integer is sufficient but a 64 bit integer is not. In domains
where large integers are needed, like cryptography and number theory,
128 bits is not enough. As for address space, 64 bits is enough for any
single computer or computer cluster. The 128 bit space is required when
we need to map everyone's stuff - a billion things for a billion people
will fill a 64 bit space. So, ultimately, 128 bits is going to be
needed. But most languages don't have a dedicated representation, so
I'll leave it out for now.

The better solution is to have a fixed size binary type parameterized by
number of bytes. Then we could write `Fixed 16` to represent `Int128`.
Integer parameters are high on my TODO list, but they aren't supported
yet. So I'll have to kick this down the road.