Evaluate general types before checking base type
Now if a type evaluates to a base type -- such as List, Tuple, Int,
Float32 -- it will be treated by the serialization engine as a base
type.

The following code, included in the new test `type-eval-1`, now works:

```
module m (foo)

-- base types
type Cpp => Int = "int"
type Cpp => List a = "std::vector<$1>" a

-- aliases at the general level
type Stack a = List a
type Vector a = List a

-- but the may be specialized at the concrete level
type Cpp => Stack a = "std::stack<$1>" a

source Cpp from "foo.hpp" ("stackToVector", "vectorToStack", "push")
stackToVector a :: Stack a -> Vector a
vectorToStack a :: Vector a -> Stack a
push a :: a -> Stack a -> Stack a

foo = stackToVector . push 42 . vectorToStack
```

Here the two `List` aliases `Stack` and `Vector` are defined. The former
is given the special `C++` definition of `std::stack`. The latter,
`Vector`, is given no special concrete meaning, so it will evaluate to
List and then to concrete `std::vector`. But both evaluate to the
general `List` type. This means both should be directly translatable to
the common binary list form (elements in sequential memory blocks). Of
course, direct support for `std::stack` will need to be added to the C++
language bindings.

The same principle could be used to support both Python `list` types AND
Python `numpy` types. This allows special high-performance handling for
directly writing numpy data to shared memory. Essentially, we can define
whatever first-class data citizens we like.

Now I have hit a new problem. I would like to use phantom types to add
annotations, like so:

```
type Cpp => L Stack a = "std::stack<$1>" a
type Cpp => L t a = List a

type L t a = List a
```

The `L` type takes a type "label" as the first argument. It may be left
generic and inferred by the typechecker. It can also be used for
concrete specialization (as shown above).

However, this is not currently allowed by the compiler. The reason is
that internally the general and concrete types are interwoven into the
`TypeF` Haskell type which requires identical forms, including the same
number of parameters. I could replace the TypeF implementation and
simply carry the general type all the way down to the code generators
where concrete inference could occur. This will require a fair bit of
refactoring and pushes more complexity into the language-specific
libraries.